refactor: refactored edit and create collection pages
