+++ client/package.json
@@ -38,16 +38,13 @@
     "clsx": "^2.1.0",
     "date-fns": "^3.3.1",
     "debounce": "^2.0.0",
-    "embla-carousel-react": "^8.0.0-rc20",
+    "embla-carousel-react": "8.0.0-rc21",
     "firebase": "^10.7.2",
-    "froala-editor": "^4.1.4",
     "i18next": "^23.7.16",
-    "js-cookie": "^3.0.5",
     "react": "^18.2.0",
     "react-auth-kit": "^3.0.2-alpha.19",
     "react-day-picker": "^8.10.0",
     "react-dom": "^18.2.0",
-    "react-froala-wysiwyg": "^4.1.4",
     "react-headroom": "^3.2.1",
     "react-hook-form": "^7.49.3",
     "react-i18next": "^14.0.0",
@@ -65,7 +62,6 @@
   },
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.10",
-    "@types/js-cookie": "^3.0.6",
     "@types/node": "^20.10.6",
     "@types/react": "^18.2.43",
     "@types/react-dom": "^18.2.17",
@@ -76,14 +72,14 @@
     "@typescript-eslint/parser": "^6.14.0",
     "@vitejs/plugin-react": "^4.2.1",
     "autoprefixer": "^10.4.16",
-    "eslint": "^8.55.0",
+    "eslint": "^8.56.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.5",
+    "eslint-plugin-unused-imports": "^3.0.0",
     "install": "^0.13.0",
     "npm": "^10.4.0",
     "postcss": "^8.4.33",
     "postcss-import": "^16.0.0",
-    "postcss-nesting": "^12.0.2",
     "prettier": "^3.1.1",
     "prettier-plugin-tailwindcss": "^0.5.11",
     "tailwind-scrollbar": "^3.0.5",
