+++ client/src/App.tsx
@@ -1,11 +1,11 @@
 import { Outlet } from '@tanstack/react-router'
-import { Header } from './components/Header'
+import { Header } from './components/ui/header/Header'
 
 export function App() {
   return (
-    <div className="font-cgr">
+    <>
       <Header />
       <Outlet />
-    </div>
+    </>
   )
 }
