+++ client/src/components/ui/authentication/AuthenticationForm.tsx
@@ -5,8 +5,8 @@ import {
   TabsTrigger,
 } from '@/components/ui/shadcn-ui/tabs'
 import { useTranslation } from 'react-i18next'
-import { LoginForm } from './ui/authentication/login-form'
-import { RegistrationForm } from './ui/authentication/registration-form'
+import { LoginForm } from './login-form'
+import { RegistrationForm } from './registration-form'
 export function AuthenticationForm() {
   const { t } = useTranslation('global')
   return (
