+++ client/src/components/ui/authentication/registration-form.tsx
@@ -12,10 +12,10 @@ import {
 } from '../shadcn-ui/form'
 import { Input } from '../shadcn-ui/input'
 import { Button } from '../shadcn-ui/button'
-import { FormError } from '../form-error'
+import { FormError } from '../forms/form-error'
 import { useRegisterMutation } from '@/api/authApi'
 import { useState } from 'react'
-import { FormSuccess } from '../form-success'
+import { FormSuccess } from '../forms/form-success'
 import useSignIn from 'react-auth-kit/hooks/useSignIn'
 import { ErrorResponse } from '@/store/reduxStore'
 import { useTranslation } from 'react-i18next'
