--- client/src/components/ui/check-auth-link.tsx
@@ -1,18 +1,0 @@
-import { useAuthStore } from '@/store/authStore'
-import { useUiStore } from '@/store/useUiStore'
-import { Link } from '@tanstack/react-router'
-
-export function CheckAuthLink({ children, className, ...props }) {
-  const { isAuth } = useAuthStore()
-  const { setIsAuthModelOpen } = useUiStore()
-
-  return (
-    <Link
-      className={className}
-      {...(isAuth ? props : null)}
-      onClick={!isAuth ? () => setIsAuthModelOpen(true) : () => {}}
-    >
-      {children}
-    </Link>
-  )
-}
