+++ client/src/components/ui/header/Header.tsx
@@ -1,13 +1,13 @@
-import { Navigation } from './ui/header/navigation'
-import SearchBar from './ui/header/search-bar'
+import { Navigation } from './navigation'
+import SearchBar from './search-bar'
 import { Link } from '@tanstack/react-router'
 import HeaderScroll from 'react-headroom'
 import { useTranslation } from 'react-i18next'
-import { SettingsButton } from './ui/header/settings-button'
-import { HamburgerMenu } from './ui/header/hamburger-menu'
-import { DialogWrapper } from './ui/dialog-wrapper'
-import { AuthenticationForm } from './AuthenticationForm'
-import { ProfileButton } from './ui/header/profile-button'
+import { SettingsButton } from './settings-button'
+import { HamburgerMenu } from './hamburger-menu'
+import { DialogWrapper } from '../dialog-wrapper'
+import { AuthenticationForm } from '../authentication/AuthenticationForm'
+import { ProfileButton } from './profile-button'
 import { useUiStore } from '@/store/useUiStore'
 import useIsAuthenticated from 'react-auth-kit/hooks/useIsAuthenticated'
