+++ client/src/components/ui/item-card.tsx
