+++ client/src/store/reduxStore.ts
@@ -31,7 +31,3 @@ export const store = configureStore({
 
 export type RootState = ReturnType<typeof store.getState>
 export type AppDispatch = typeof store.dispatch
-export type ErrorResponse = {
-  status: string
-  data: string
-}
