+++ client/src/translations/ru/global.json
@@ -23,6 +23,14 @@
   "en": "Английский",
   "ru": "Русский",
   "forms": {
+    "create": {
+      "success": "Коллекция добавлена!",
+      "submit": "Создать новую коллекцию!"
+    },
+    "update": {
+      "success": "Коллекция изменена!",
+      "submit": "Обновить коллекцию!"
+    },
     "successLogin": "Авторизовано успешно!",
     "successRegister": "Успешная регистрация!",
     "password": "Пароль",
@@ -30,8 +38,6 @@
     "register": "Регистрация",
     "tagsAreRequired": "Теги обязательны",
     "itemAdded": "Предмет добавлен!",
-    "collectionAdded": "Коллекция добавлена!",
-    "collectionEdited": "Коллекция изменена!",
     "image": "Картинка",
     "name": "Название",
     "itemName": "Название предмета",
@@ -47,12 +53,10 @@
     "collectionName": "Название коллекции",
     "author": "Автор",
     "closedMode": "Приватный режим",
-    "createCollection": "Создать новую коллекцию!",
     "startWriting": "Начинай писать :0",
     "cancel": "Отмена",
     "delete": "Удалить",
-    "apply": "Применить",
-    "updateCollection": "Обновить коллекцию!"
+    "apply": "Применить"
   },
   "home": {
     "five": "Пять самых больших коллекций",
